excluir evento funcionando

+++ controller/controllerEditarEvento.php
@@ -12,25 +12,44 @@ function editarEvento(){
     $registro = mysqli_fetch_array($busca);
     $arquivo = $registro['imagem'];
 
-    if(getOpcao()) {
-        if (isset($_FILES['arquivo'])) {
-            unlink("../uploads/"."$arquivo");//deletando a imagem antiga do servidor
-            $extensao = strtolower(substr($_FILES['arquivo']['name'], -4)); //pega a extensão do arquivo
-            if ($extensao == 'jpeg') {
-                $novoNome = md5(time()) . '.' . $extensao; // define o novo nome do arquivo
+    if (getEscolha()) {
+        if (getOpcao()) {
+            if (isset($_FILES['arquivo'])) {
+                unlink("../uploads/" . "$arquivo");//deletando a imagem antiga do servidor
+                $extensao = strtolower(substr($_FILES['arquivo']['name'], -4)); //pega a extensão do arquivo
+                if ($extensao == 'jpeg') {
+                    $novoNome = md5(time()) . '.' . $extensao; // define o novo nome do arquivo
+                } else {
+                    $novoNome = md5(time()) . $extensao; // define o novo nome do arquivo
+                }
+
+                $diretorio = "../uploads/"; //define o diretorio para onde será enviado o arquivo
+
+                move_uploaded_file($_FILES['arquivo']['tmp_name'], $diretorio . $novoNome);//efetua o upload
+
+                $data = getData() . " " . getHora() . ":00"; // concatenando a data
+                $nome = getNome();
+                $descricao = getDescricao();
+                $query = mysqli_query($conecta, "UPDATE evento
+                                                  SET data = '$data', imagem = '$novoNome', descricao = '$descricao'
+                                                  WHERE nome = '$nome'");
+                if ($query) {
+                    echo 'Salvo com sucesso!';
+                    header('Location: ../index.php');
+                } else {
+                    echo 'Não foi possivel salvar a edição.';
+                    echo mysqli_error($conecta);
+                }
+
             } else {
-                $novoNome = md5(time()) . $extensao; // define o novo nome do arquivo
+                echo '<p> arquivo não selecionado</p>';
             }
-
-            $diretorio = "../uploads/"; //define o diretorio para onde será enviado o arquivo
-
-            move_uploaded_file($_FILES['arquivo']['tmp_name'], $diretorio . $novoNome);//efetua o upload
-
+        } else {
             $data = getData() . " " . getHora() . ":00"; // concatenando a data
             $nome = getNome();
             $descricao = getDescricao();
             $query = mysqli_query($conecta, "UPDATE evento
-                                                  SET data = '$data', imagem = '$novoNome', descricao = '$descricao'
+                                                  SET data = '$data', descricao = '$descricao'
                                                   WHERE nome = '$nome'");
             if ($query) {
                 echo 'Salvo com sucesso!';
@@ -40,25 +59,10 @@ function editarEvento(){
                 echo mysqli_error($conecta);
             }
 
-        } else {
-            echo '<p> arquivo não selecionado</p>';
         }
     }
     else{
-        $data = getData() . " " . getHora() . ":00"; // concatenando a data
-        $nome = getNome();
-        $descricao = getDescricao();
-        $query = mysqli_query($conecta, "UPDATE evento
-                                                  SET data = '$data', descricao = '$descricao'
-                                                  WHERE nome = '$nome'");
-        if ($query) {
-            echo 'Salvo com sucesso!';
-            header('Location: ../index.php');
-        } else {
-            echo 'Não foi possivel salvar a edição.';
-            echo mysqli_error($conecta);
-        }
-
+        ExcluirEvento($nome, $arquivo);
     }
 }
 
@@ -126,5 +130,23 @@ function getOpcao(){
     else {
         return false;
     }
+}
+
+function getEscolha(){
+    if ($_POST["escolha"] == "Salvar"){
+        return true;
+    }
+    else {
+        return false;
+    }
+
+}
+
+function ExcluirEvento ($nome,$arquivo){
+    $conecta = mysqli_connect('localhost', 'root', '');
+    mysqli_select_db($conecta, 'database');
+    mysqli_query($conecta,"DELETE FROM evento WHERE nome = '$nome'");
+    unlink("../uploads/" . "$arquivo");//deletando a imagem antiga do servidor
+    header('Location: ../index.php');
 }
 ?>
